+++ tests/Rest/RemoveAllFormSubmissionsControllerTest.php
@@ -6,6 +6,7 @@ namespace Dbp\Relay\FormalizeBundle\Tests\Rest;
 
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Dbp\Relay\FormalizeBundle\Authorization\AuthorizationService;
+use Dbp\Relay\FormalizeBundle\Entity\Submission;
 use Dbp\Relay\FormalizeBundle\Rest\RemoveAllFormSubmissionsController;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -39,6 +40,75 @@ class RemoveAllFormSubmissionsControllerTest extends RestTestCase
 
         $this->assertNull($this->getSubmission($submission->getIdentifier()));
         $this->assertNull($this->getSubmission($submission2->getIdentifier()));
+    }
+
+    public function testRemoveAllFormSubmissionsGrantBasedSubmissionAuthorization(): void
+    {
+        $form = $this->addForm(grantBasedSubmissionAuthorization: true);
+        $submission = $this->addSubmission($form);
+        $submission2 = $this->addSubmission($form);
+
+        $this->assertNotNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submission2->getIdentifier()));
+
+        $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
+            AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
+            AuthorizationService::DELETE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
+
+        $this->removeAllFormSubmissionsController->__invoke(self::createRequestStack(
+            '/formalize/submissions?formIdentifier='.$form->getIdentifier(), 'DELETE')->getCurrentRequest());
+
+        $this->assertNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNull($this->getSubmission($submission2->getIdentifier()));
+    }
+
+    public function testRemoveAllFormSubmissionsExceptForDrafts(): void
+    {
+        $form = $this->addForm(
+            allowedSubmissionStates: Submission::SUBMISSION_STATE_DRAFT | Submission::SUBMISSION_STATE_SUBMITTED);
+        $submission = $this->addSubmission($form);
+        $submissionDraft = $this->addSubmission($form, submissionState: Submission::SUBMISSION_STATE_DRAFT);
+        $submission2 = $this->addSubmission($form);
+
+        $this->assertNotNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submissionDraft->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submission2->getIdentifier()));
+
+        $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
+            AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
+            AuthorizationService::DELETE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
+
+        $this->removeAllFormSubmissionsController->__invoke(self::createRequestStack(
+            '/formalize/submissions?formIdentifier='.$form->getIdentifier(), 'DELETE')->getCurrentRequest());
+
+        $this->assertNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNull($this->getSubmission($submission2->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submissionDraft->getIdentifier()));
+    }
+
+    public function testRemoveAllFormSubmissionsExceptForDraftsGrantBasedAuthorization(): void
+    {
+        $form = $this->addForm(
+            grantBasedSubmissionAuthorization: true,
+            allowedSubmissionStates: Submission::SUBMISSION_STATE_DRAFT | Submission::SUBMISSION_STATE_SUBMITTED);
+        $submission = $this->addSubmission($form);
+        $submissionDraft = $this->addSubmission($form, submissionState: Submission::SUBMISSION_STATE_DRAFT);
+        $submission2 = $this->addSubmission($form);
+
+        $this->assertNotNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submissionDraft->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submission2->getIdentifier()));
+
+        $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
+            AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
+            AuthorizationService::DELETE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
+
+        $this->removeAllFormSubmissionsController->__invoke(self::createRequestStack(
+            '/formalize/submissions?formIdentifier='.$form->getIdentifier(), 'DELETE')->getCurrentRequest());
+
+        $this->assertNull($this->getSubmission($submission->getIdentifier()));
+        $this->assertNull($this->getSubmission($submission2->getIdentifier()));
+        $this->assertNotNull($this->getSubmission($submissionDraft->getIdentifier()));
     }
 
     public function testRemoveAllFormSubmissionsForbidden(): void